Change datatype of sildeShowDelay to double

+++ Phototonic.cpp
@@ -811,7 +811,7 @@ void Phototonic::createActions() {
     connect(resetZoomAction, &QAction::triggered, this, [=](){
         imageViewer->zoomTo(imageViewer->zoomMode() == ImageViewer::ZoomToFit ?
                                                                     ImageViewer::ZoomToFill :
-                                                                    ImageViewer::ZoomToFit); 
+                                                                    ImageViewer::ZoomToFit);
     });
 
     MAKE_ACTION(origZoomAction, tr("Original Size"), "origZoom", "/");
@@ -1895,7 +1895,7 @@ void Phototonic::batchTransform() {
                 MessageBox msgBox(this);
                 msgBox.critical(tr("Error"), tr("Failed to copy or move image."));
                 return;
-            } 
+            }
         }
     }
 
@@ -2445,7 +2445,7 @@ void Phototonic::readSettings() {
     Settings::showHiddenFiles = Settings::value(Settings::optionShowHiddenFiles, false).toBool();
     Settings::wrapImageList = Settings::value(Settings::optionWrapImageList, false).toBool();
     Settings::defaultSaveQuality = Settings::value(Settings::optionDefaultSaveQuality, 90).toInt();
-    Settings::slideShowDelay = Settings::value(Settings::optionSlideShowDelay, 5).toInt();
+    Settings::slideShowDelay = Settings::value(Settings::optionSlideShowDelay, 5.0).toDouble();
     Settings::slideShowRandom = Settings::value(Settings::optionSlideShowRandom, false).toBool();
     Settings::showImageName = Settings::value(Settings::optionShowImageName, false).toBool();
     Settings::smallToolbarIcons = Settings::value(Settings::optionSmallToolbarIcons, true).toBool();
@@ -2745,7 +2745,7 @@ void Phototonic::toggleSlideShow() {
 
         SlideShowTimer = new QTimer(this);
         connect(SlideShowTimer, SIGNAL(timeout()), this, SLOT(slideShowHandler()));
-        SlideShowTimer->start(Settings::slideShowDelay * 1000);
+        SlideShowTimer->start(Settings::slideShowDelay * 1000.0);
 
         slideShowAction->setText(tr("Stop Slide Show"));
         imageViewer->setFeedback(tr("Slide show started"));
@@ -3091,7 +3091,7 @@ void Phototonic::reloadThumbs() {
     m_imageTags->removeTransientTags();
 
     if (findDupesAction->isChecked()) {
-        const bool actionEnabled[5] = { goBackAction->isEnabled(), goFrwdAction->isEnabled(), 
+        const bool actionEnabled[5] = { goBackAction->isEnabled(), goFrwdAction->isEnabled(),
                                         goUpAction->isEnabled(), goHomeAction->isEnabled(), refreshAction->isEnabled() };
         goBackAction->setEnabled(false);
         goFrwdAction->setEnabled(false);
@@ -3433,7 +3433,7 @@ bool Phototonic::eventFilter(QObject *o, QEvent *e)
                 animator->setEasingCurve(QEasingCurve::InOutQuad);
             }
             const int grid = thumbsViewer->gridSize().height();
-            int v = (animator->state() == QAbstractAnimation::Running) ? animator->endValue().toInt() : 
+            int v = (animator->state() == QAbstractAnimation::Running) ? animator->endValue().toInt() :
                                                                          thumbsViewer->verticalScrollBar()->value();
             animator->setStartValue(v);
             if (qAbs(steps) == 1000) {
